Refactor DFP Data Construction
